Refactors the M&C testbed
