+++ testbeds/testbed-mc/src/main/java/org/ccsds/mo/mc/testbed/backends/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Package containing the MPD testbed backends. These will be used by the testing classes.
  */
-package org.ccsds.mo.mpd.testbed.backends;
+package org.ccsds.mo.mc.testbed.backends;
