+++ testbeds/testbed-mc/src/main/java/org/ccsds/mo/mc/testbed/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Package containing utility classes for the MC testbed.
  */
-package org.ccsds.mo.mpd.testbed;
+package org.ccsds.mo.mc.testbed;
