+++ testbeds/testbed-mc/src/test/java/org/ccsds/mo/mc/testbed/UC1_Ex1_Test.java
@@ -18,7 +18,7 @@
  * limitations under the License. 
  * ----------------------------------------------------------------------------
  */
-package org.ccsds.mo.mpd.testbed;
+package org.ccsds.mo.mc.testbed;
 
 import java.io.IOException;
 import org.ccsds.moims.mo.mal.structures.UInteger;
